Add quality assurance section to quality control lesson
There are two stages to making sure that the data entered in a dataset is correct.
The first of these  is quality assurance, which are approaches that prevent bad data
from ever getting into the dataset in the first place. This change briefly explains
the different between quality assurance and control and then shows how to implement
quality assurance in spreadsheets.